Title: 체질량지수, Time: 91ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 240609/체질량지수/bmi.java
@@ -14,7 +14,7 @@ public class Main {
 
     b = 10_000.0f * w / h / h;
 
-    System.out.printf("%d\n", b);
+    System.out.printf("%d\n", (int)b);
     if(b >= 25)
         System.out.println("Obesty");
   } //  main-end
